pyproject.toml
[project]
name = "project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.53.1",
    "torchvision>=0.25.0",
    "ultralytics>=8.4.6",
]
